refactor: script decoded
